Feature/issue 7.1

+++ CrossReview/src/CrossReview.Application/Template/UseCases/AddQuestionToTemplate/AddQuestionRequest.cs
@@ -1,3 +1,3 @@
 ﻿namespace CrossReview.Application.Template.UseCases.AddQuestionToTemplate;
 
-public record AddQuestionRequest(Guid TemplateId, string Title, int Weight);
+public record AddQuestionRequest(Guid TemplateId, string Title, double Weight);
